datos aemet

+++ sql/AGRARIO.sql
@@ -1,10 +1,10 @@
 create table AGRARIO (
   MUN TEXT,
   YR INTEGER,
-  "sau" INTEGER,
-"uta" INTEGER,
-"explotaciones" INTEGER,
-"unidadesganaderas" INTEGER,
+  "sau" REAL,
+  "uta" REAL,
+  "explotaciones" INTEGER,
+  "unidadesganaderas" REAL,
   PRIMARY KEY (MUN, YR),
   FOREIGN KEY(MUN) REFERENCES municipios(ID)
 )
