+++ sql/RENTA.sql
@@ -2,8 +2,8 @@ create table RENTA (
   MUN TEXT,
   YR INTEGER,
   "declaraciones" INTEGER,
-"renta" INTEGER,
-"tipo" INTEGER,
+  "renta" REAL,
+  "tipo" INTEGER,
   PRIMARY KEY (MUN, YR),
   FOREIGN KEY(MUN) REFERENCES municipios(ID)
 )
